🚚 move unit tests into prod modules
