--- marketplace-domain/src/contribution/aggregate_root/tests/mod.rs
@@ -1,3 +1,0 @@
-mod contribution;
-mod event;
-mod status;
